refactor!: observer should imply implementation of validationRules() method
ValidateModel observer now requires observed models to implement the
SelfValidatingModel interface.

BREAKING CHANGE: models without validationRules() method will cause
fatal error.

Resolves #14
