+++ tests/Fixtures/Models/NotSelfValidatingCar.php
@@ -15,36 +15,40 @@
 // You should have received a copy of the GNU General Public License along with
 // package spoorsny/laravel-model-validating-observer. If not, see <https://www.gnu.org/licenses/>.
 
-use Illuminate\Database\Migrations\Migration;
-use Illuminate\Database\Schema\Blueprint;
-use Illuminate\Support\Facades\Schema;
+namespace Spoorsny\Laravel\Tests\Fixtures\Models;
+
+use Illuminate\Database\Eloquent\Attributes\ObservedBy;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+
+use Spoorsny\Laravel\Observers\ValidateModel;
 
 /**
- * Create database table used during testing.
- *
  * @author     Geoffrey Bernardo van Wyk <geoffrey@vanwyk.biz>
  * @copyright  2024 Geoffrey Bernardo van Wyk {@link https://geoffreyvanwyk.dev}
  * @license    {@link http://www.gnu.org/copyleft/gpl.html} GNU GPL v3 or later
  */
-return new class () extends Migration {
-    /**
-     * Run the migrations.
-     */
-    public function up(): void
-    {
-        Schema::create('without_validation_rules', function (Blueprint $table) {
-            $table->id();
-            $table->string('make');
-            $table->string('model');
-            $table->timestamps();
-        });
-    }
+#[ObservedBy(ValidateModel::class)]
+class NotSelfValidatingCar extends Model
+{
+    use HasFactory;
 
     /**
-     * Reverse the migrations.
+     * Rules for validating the model's attributes.
+     *
+     * @see     {@link https://laravel.com/docs/11.x/validation#available-validation-rules}
+     * @return  array<string,array<string,mixed>>
      */
-    public function down(): void
+    public static function validationRules(): array
     {
-        Schema::dropIfExists('without_validation_rules');
+        return [
+            'rules' => [
+                'make' => 'required|string',
+                'model' => 'required|string',
+            ],
+            'messages' => [
+                'make.required' => 'We need to know the make of the car.',
+            ],
+        ];
     }
-};
+}
